+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v4.3.1] 2021-02-18
+
+### Change
+
+- Command config:oxid62:modules-configuration filter exists module setting
+- Command config:oxid62:modules-configuration save variable type
+
 ## [v4.3.0] 2021-02-18
 
 ### Add